requirements.txt
# AI Chatbot - Local, Free, with Emotions & Character Face
# Python 3.8+ required

# Local LLM (Ollama)
ollama>=0.3.0

# Web UI
gradio>=4.0.0

# Sentiment Analysis (runs locally, free)
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0

# Text-to-Speech (offline, high quality)
edge-tts>=6.1.0

# Speech Recognition (voice input)
SpeechRecognition>=3.10.0
pyaudio>=0.2.11

# Audio processing for lip sync
numpy>=1.24.0
scipy>=1.10.0

# Image handling
Pillow>=10.0.0

# Video processing for talking head
opencv-python>=4.8.0
moviepy>=1.0.3

# Wav2Lip dependencies
librosa>=0.10.0
numba>=0.56.0
